devnets: break glamsterdam analysis into subtabs
