+++ src/components/glamsterdam/EipCandidatesTab.tsx
@@ -1,0 +1,1 @@
+export { default } from '../devnets/GlamsterdamPrioritizationSection';
