+++ src/data/devnets/bal-devnet-5.json
@@ -1,0 +1,272 @@
+{
+  "id": "bal-devnet-5",
+  "title": "bal-devnet-5",
+  "sourceUrl": "https://notes.ethereum.org/@ethpandaops/bal-devnet-5",
+  "scrapedAt": "2026-05-06T13:20:57.863Z",
+  "announcements": [
+    "bal-devnet-5 will launch on wednesday **29.04**.",
+    "BAL benchmark dashboard https://nerolation.github.io/bal-dashboard/. Based on data from https://benchmarkoor.core.ethpandaops.io. Requires benchmarkoor api key.",
+    "EIP-8037: bal-devnet-5 keeps the **static** `cost_per_state_byte = 1174`. It also includes the changes new accounting changes at the frame boundry. see https://github.com/ethereum/EIPs/pull/11573"
+  ],
+  "eips": [
+    {
+      "number": 7708,
+      "title": "ETH transfers emit a log — now includes `CREATE`/`CREATE2`",
+      "status": "updated",
+      "url": "https://eips.ethereum.org/EIPS/eip-7708"
+    },
+    {
+      "number": 7778,
+      "title": "Block Gas Accounting without Refunds",
+      "status": null,
+      "url": "https://eips.ethereum.org/EIPS/eip-7778"
+    },
+    {
+      "number": 7843,
+      "title": "SLOTNUM opcode",
+      "status": null,
+      "url": "https://eips.ethereum.org/EIPS/eip-7843"
+    },
+    {
+      "number": 7928,
+      "title": "Block-Level Access Lists — block access index `uint32`",
+      "status": "updated",
+      "url": "https://eips.ethereum.org/EIPS/eip-7928"
+    },
+    {
+      "number": 7954,
+      "title": "Increase Maximum Contract Size",
+      "status": null,
+      "url": "https://eips.ethereum.org/EIPS/eip-7954"
+    },
+    {
+      "number": 7975,
+      "title": "eth/70 – partial block receipt lists",
+      "status": "optional",
+      "url": "https://eips.ethereum.org/EIPS/eip-7975"
+    },
+    {
+      "number": 8024,
+      "title": "Backward compatible SWAPN, DUPN, EXCHANGE",
+      "status": null,
+      "url": "https://eips.ethereum.org/EIPS/eip-8024"
+    },
+    {
+      "number": 8037,
+      "title": "State Creation Gas Cost Increase — ~~dynamic `cpsb`~~ static `cpsb=1174` and updated accounting see https://github.com/ethereum/EIPs/pull/11573",
+      "status": "updated",
+      "url": "https://eips.ethereum.org/EIPS/eip-8037"
+    },
+    {
+      "number": 8159,
+      "title": "eth/71 – Block Access List Exchange — empty BAL response `0x80`",
+      "status": "updated",
+      "url": "https://eips.ethereum.org/EIPS/eip-8159"
+    },
+    {
+      "number": 7976,
+      "title": "Increase Calldata Floor Cost (64/64)",
+      "status": "new",
+      "url": "https://eips.ethereum.org/EIPS/eip-7976"
+    },
+    {
+      "number": 7981,
+      "title": "Increase Access List Cost",
+      "status": "new",
+      "url": "https://eips.ethereum.org/EIPS/eip-7981"
+    }
+  ],
+  "elClientSupport": {
+    "clients": [
+      "Geth",
+      "Besu",
+      "Reth",
+      "Nethermind",
+      "Erigon",
+      "Nimbus-EL",
+      "Ethrex"
+    ],
+    "matrix": [
+      {
+        "eipNumber": 7708,
+        "label": "7708 (ETH Logs)",
+        "support": {
+          "Geth": "supported",
+          "Besu": "supported",
+          "Reth": "supported",
+          "Nethermind": "supported",
+          "Erigon": "supported",
+          "Nimbus-EL": "supported",
+          "Ethrex": "supported"
+        }
+      },
+      {
+        "eipNumber": 7778,
+        "label": "7778 (Gas Refunds)",
+        "support": {
+          "Geth": "supported",
+          "Besu": "supported",
+          "Reth": "supported",
+          "Nethermind": "supported",
+          "Erigon": "supported",
+          "Nimbus-EL": "supported",
+          "Ethrex": "supported"
+        }
+      },
+      {
+        "eipNumber": 7843,
+        "label": "7843 (SLOTNUM)",
+        "support": {
+          "Geth": "supported",
+          "Besu": "supported",
+          "Reth": "supported",
+          "Nethermind": "supported",
+          "Erigon": "supported",
+          "Nimbus-EL": "supported",
+          "Ethrex": "supported"
+        }
+      },
+      {
+        "eipNumber": 7928,
+        "label": "7928 (BAL)",
+        "support": {
+          "Geth": "supported",
+          "Besu": "supported",
+          "Reth": "supported",
+          "Nethermind": "supported",
+          "Erigon": "supported",
+          "Nimbus-EL": "supported",
+          "Ethrex": "supported"
+        }
+      },
+      {
+        "eipNumber": 7954,
+        "label": "7954 (MContract)",
+        "support": {
+          "Geth": "supported",
+          "Besu": "supported",
+          "Reth": "supported",
+          "Nethermind": "supported",
+          "Erigon": "supported",
+          "Nimbus-EL": "supported",
+          "Ethrex": "supported"
+        }
+      },
+      {
+        "eipNumber": 7975,
+        "label": "7975 (eth/70)",
+        "support": {
+          "Geth": "supported",
+          "Besu": "supported",
+          "Reth": "supported",
+          "Nethermind": "supported",
+          "Erigon": "supported",
+          "Nimbus-EL": "supported",
+          "Ethrex": "supported"
+        }
+      },
+      {
+        "eipNumber": 8024,
+        "label": "8024 (SWAPN/DUPN)",
+        "support": {
+          "Geth": "supported",
+          "Besu": "supported",
+          "Reth": "supported",
+          "Nethermind": "supported",
+          "Erigon": "supported",
+          "Nimbus-EL": "supported",
+          "Ethrex": "supported"
+        }
+      },
+      {
+        "eipNumber": 8037,
+        "label": "8037 (+frame accounting)",
+        "support": {
+          "Geth": "in_progress",
+          "Besu": "in_progress",
+          "Reth": "in_progress",
+          "Nethermind": "in_progress",
+          "Erigon": "in_progress",
+          "Nimbus-EL": "in_progress",
+          "Ethrex": "in_progress"
+        }
+      },
+      {
+        "eipNumber": 7976,
+        "label": "7976 (Calldata)",
+        "support": {
+          "Geth": "in_progress",
+          "Besu": "in_progress",
+          "Reth": "in_progress",
+          "Nethermind": "in_progress",
+          "Erigon": "in_progress",
+          "Nimbus-EL": "in_progress",
+          "Ethrex": "in_progress"
+        }
+      },
+      {
+        "eipNumber": 7981,
+        "label": "7981 (AL Cost)",
+        "support": {
+          "Geth": "in_progress",
+          "Besu": "in_progress",
+          "Reth": "in_progress",
+          "Nethermind": "in_progress",
+          "Erigon": "in_progress",
+          "Nimbus-EL": "in_progress",
+          "Ethrex": "in_progress"
+        }
+      },
+      {
+        "eipNumber": 8159,
+        "label": "8159 (eth/71)",
+        "support": {
+          "Geth": "not_supported",
+          "Besu": "supported",
+          "Reth": "supported",
+          "Nethermind": "not_supported",
+          "Erigon": "not_supported",
+          "Nimbus-EL": "supported",
+          "Ethrex": "in_progress"
+        }
+      }
+    ]
+  },
+  "clClientSupport": {
+    "clients": [
+      "Lodestar",
+      "Lighthouse",
+      "Prysm"
+    ],
+    "matrix": [
+      {
+        "eipNumber": 7928,
+        "label": "EIP-7928",
+        "support": {
+          "Lodestar": "supported",
+          "Lighthouse": "supported",
+          "Prysm": "bal-devnet-1"
+        }
+      },
+      {
+        "eipNumber": 7843,
+        "label": "EIP-7843",
+        "support": {
+          "Lodestar": "supported",
+          "Lighthouse": "supported",
+          "Prysm": "not_supported"
+        }
+      }
+    ]
+  },
+  "specReferences": {
+    "consensusSpecs": {
+      "version": "v1.6.1",
+      "url": "https://github.com/ethereum/consensus-specs/releases/tag/v1.6.1"
+    },
+    "executionSpecs": {
+      "version": "snøbal-devnet-5@v1.0.0",
+      "url": "https://github.com/ethereum/execution-spec-tests/releases/tag/snøbal-devnet-5@v1.0.0"
+    }
+  }
+}
