+++ src/data/devnets/epbs-devnet-2.json
@@ -1,0 +1,43 @@
+{
+  "id": "epbs-devnet-2",
+  "title": "epbs-devnet-2",
+  "sourceUrl": "https://notes.ethereum.org/@ethpandaops/epbs-devnet-2",
+  "scrapedAt": "2026-05-06T13:20:59.947Z",
+  "announcements": [
+    "epbs-devnet-2 will no longer happen. For latest info please head over to https://notes.ethereum.org/@ethpandaops/glamsterdam-devnet-0",
+    "Everything below is outdated!!"
+  ],
+  "eips": [],
+  "elClientSupport": {
+    "clients": [],
+    "matrix": []
+  },
+  "clClientSupport": {
+    "clients": [
+      "Lodestar",
+      "Lighthouse",
+      "Prysm",
+      "Teku",
+      "Grandine",
+      "Nimbus"
+    ],
+    "matrix": [
+      {
+        "eipNumber": 7732,
+        "label": "EIP-7732",
+        "support": {
+          "Lodestar": "unknown",
+          "Lighthouse": "unknown",
+          "Prysm": "unknown",
+          "Teku": "unknown",
+          "Grandine": "unknown",
+          "Nimbus": "unknown"
+        }
+      }
+    ]
+  },
+  "specReferences": {
+    "consensusSpecs": null,
+    "executionSpecs": null
+  }
+}
