+++ src/data/devnets/glamsterdam.json
@@ -1,6 +1,6 @@
 {
   "upgrade": "glamsterdam",
-  "lastUpdated": "2026-04-16",
+  "lastUpdated": "2026-05-06",
   "devnets": [
     {
       "id": "bal-devnet-0",
@@ -35,14 +35,6 @@
       "eips": [7732]
     },
     {
-      "id": "epbs-devnet-1",
-      "type": "headliner",
-      "headliner": "EPBS",
-      "version": 1,
-      "launchDate": "2026-03-31",
-      "eips": [7732]
-    },
-    {
       "id": "bal-devnet-3",
       "type": "headliner",
       "headliner": "BAL",
@@ -53,6 +45,14 @@
       "optionalEips": [7975, 8159]
     },
     {
+      "id": "epbs-devnet-1",
+      "type": "headliner",
+      "headliner": "EPBS",
+      "version": 1,
+      "launchDate": "2026-03-31",
+      "eips": [7732]
+    },
+    {
       "id": "bal-devnet-4",
       "type": "headliner",
       "headliner": "BAL",
@@ -61,6 +61,42 @@
       "isTarget": true,
       "eips": [7928, 8024, 7843, 7708, 7778, 8037, 7954, 7976, 7981],
       "updatedEips": [8037],
+      "optionalEips": [7975, 8159]
+    },
+    {
+      "id": "glamsterdam-devnet-0",
+      "type": "combined",
+      "version": 0,
+      "launchDate": "2026-04-24",
+      "eips": [7708, 7732, 7778, 7843, 7928, 7954, 8024, 7976, 7981, 8159],
+      "updatedEips": [7732, 7928, 8159],
+      "optionalEips": [7975]
+    },
+    {
+      "id": "bal-devnet-5",
+      "type": "headliner",
+      "headliner": "BAL",
+      "version": 5,
+      "launchDate": "2026-04-29",
+      "eips": [7928, 8024, 7843, 7708, 7778, 8037, 7954, 7976, 7981],
+      "optionalEips": [7975, 8159]
+    },
+    {
+      "id": "glamsterdam-devnet-2",
+      "type": "combined",
+      "version": 2,
+      "launchDate": "2026-04-30",
+      "eips": [7708, 7732, 7778, 7843, 7928, 7954, 8024, 7976, 7981, 8061],
+      "updatedEips": [8061],
+      "optionalEips": [7975, 8159]
+    },
+    {
+      "id": "bal-devnet-6",
+      "type": "headliner",
+      "headliner": "BAL",
+      "version": 6,
+      "launchDate": "2026-05-01",
+      "eips": [7928, 8024, 7843, 7708, 7778, 8037, 7954, 7976, 7981],
       "optionalEips": [7975, 8159]
     }
   ]
